Frames: new euclidian sequencer mode.
(reordered the modes using the big knob for Keyframe mode)

This is a quad euclidean sequencer mode: for each channel we can set independently:
- fill ratio of the pattern (with the knobs)
- length of the pattern (long-pressing ADD, defaults to 16)
- shape of the envelope at the output (long-pressing DEL, defaults to small ping)
The big knob sets the amount of rotation of each channel. The Frame CV reads the clock
that makes the sequencer step. The Fr.Step output is high when the XOR of all channels
is negative (as indicated by the frame LED).